+++ src/edu/smu/cse7343/bakos/os/Memory.java
@@ -9,10 +9,10 @@ package edu.smu.cse7343.bakos.os;
 
 public class Memory {
 
-    int[] registers;
+    float[] registers;
 
     public Memory(int size) {
-        registers = new int[size];
+        registers = new float[size];
     }
 
     public int totalSize() {
